There are different ways to solve a problem
As a good practice I intent to write my code all at once, and when it works
the next step is cleaning the code, remove unnecessary parts, and make
it more readable not just for me.

One of my goals is trying to re-do the same problem with recursion, but
not all of them, only with the problems that I like the most.